qa: Improve usage with curl
Prior to this commit, the payloads of the store and retrieve requests
had fields that were []byte, which made it sligthly difficult to create
instructions on the usage using curl.

This commit alters these fields and base64 encodes/decodes the key field
so that the flow involves less use of commands for encoding strings to
byte arrays within a shell.